eslint limpo e dependencia ciclica ok e mudança na estrutura das rotas e persistencia
