+++ src/index.html
@@ -24,7 +24,7 @@
   <script src="/__/firebase/8.7.1/firebase-firestore.js"></script>
   <script src="/__/firebase/8.7.1/firebase-storage.js"></script>
   <script src="/__/firebase/init.js?useEmulator=true"></script>
-  <script type="module" src="router.js"></script>
+  <script type="module" src="main.js"></script>
 
 </body>
