+++ src/main.js
@@ -1,0 +1,4 @@
+import { routeRender } from './routes/render.js';
+
+window.addEventListener('popstate', routeRender);
+window.addEventListener('load', routeRender);
