+++ src/routes/navigation.js
@@ -1,0 +1,5 @@
+export function navigation(path) {
+  window.history.pushState({}, '', path);
+  const popStateEvent = new PopStateEvent('popstate', { state: {} });
+  dispatchEvent(popStateEvent);
+}
